refactor: config codes organization
