+++ crates/mako/src/config/inline_css.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+use crate::create_deserialize_fn;
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct InlineCssConfig {}
+
+create_deserialize_fn!(deserialize_inline_css, InlineCssConfig);
