+++ crates/mako/src/config/mako.config.default.json
@@ -1,0 +1,79 @@
+{
+  "entry": {},
+  "output": {
+    "path": "dist",
+    "mode": "bundle",
+    "esVersion": "es2022",
+    "meta": false,
+    "chunkLoadingGlobal": "",
+    "preserveModules": false,
+    "preserveModulesRoot": "",
+    "skipWrite": false,
+    "crossOriginLoading": false,
+    "globalModuleRegistry": false
+  },
+  "resolve": { "alias": [], "extensions": ["js", "jsx", "ts", "tsx"] },
+  "mode": "development",
+  "minify": true,
+  "devtool": "source-map",
+  "externals": {},
+  "copy": ["public"],
+  "providers": {},
+  "publicPath": "/",
+  "inlineLimit": 10000,
+  "inlineExcludesExtensions": [],
+  "targets": { "chrome": 80 },
+  "less": { "theme": {}, "lesscPath": "", "javascriptEnabled": true },
+  "define": {},
+  "mdx": false,
+  "platform": "browser",
+  "hmr": {},
+  "moduleIdStrategy": "named",
+  "hash": false,
+  "_treeShaking": "basic",
+  "autoCSSModules": false,
+  "ignoreCSSParserErrors": false,
+  "dynamicImportToRequire": false,
+  "writeToDisk": true,
+  "transformImport": [],
+  "chunkParallel": true,
+  "clean": true,
+  "nodePolyfill": true,
+  "ignores": [],
+  "optimizePackageImports": false,
+  "emotion": false,
+  "flexBugs": false,
+  "cjs": false,
+  "optimization": { "skipModules": true, "concatenateModules": true },
+  "react": {
+    "pragma": "React.createElement",
+    "importSource": "react",
+    "runtime": "automatic",
+    "pragmaFrag": "React.Fragment"
+  },
+  "progress": {
+    "progressChars": "▨▨"
+  },
+  "duplicatePackageChecker": {
+    "verbose": false,
+    "showHelp": false,
+    "emitError": false
+  },
+  "emitAssets": true,
+  "cssModulesExportOnlyLocales": false,
+  "inlineCSS": false,
+  "rscServer": false,
+  "rscClient": false,
+  "experimental": {
+    "webpackSyntaxValidate": [],
+    "requireContext": true,
+    "detectCircularDependence": {
+      "ignores": ["node_modules"],
+      "graphviz": false
+    }
+  },
+  "useDefineForClassFields": true,
+  "emitDecoratorMetadata": false,
+  "watch": { "ignorePaths": [], "_nodeModulesRegexes": [] },
+  "devServer": { "host": "127.0.0.1", "port": 3000 }
+}
