+++ crates/mako/src/config/mode.rs
@@ -1,0 +1,16 @@
+use clap::ValueEnum;
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug, PartialEq, Eq, ValueEnum, Clone)]
+pub enum Mode {
+    #[serde(rename = "development")]
+    Development,
+    #[serde(rename = "production")]
+    Production,
+}
+
+impl std::fmt::Display for Mode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.to_possible_value().unwrap().get_name().fmt(f)
+    }
+}
