+++ crates/mako/src/config/module_id_strategy.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Clone, Copy, Debug)]
+pub enum ModuleIdStrategy {
+    #[serde(rename = "hashed")]
+    Hashed,
+    #[serde(rename = "named")]
+    Named,
+    #[serde(rename = "numeric")]
+    Numeric,
+}
