+++ crates/mako/src/config/provider.rs
@@ -1,0 +1,7 @@
+use std::collections::HashMap;
+
+// format: HashMap<identifier, (import_source, specifier)>
+// e.g.
+// { "process": ("process", "") }
+// { "Buffer": ("buffer", "Buffer") }
+pub type Providers = HashMap<String, (String, String)>;
