+++ crates/mako/src/utils.rs
@@ -7,6 +7,8 @@ pub(crate) mod test_helper;
 pub mod thread_pool;
 pub mod tokio_runtime;
 
+use std::path::Path;
+
 use anyhow::{anyhow, Result};
 use base64::engine::general_purpose;
 use base64::Engine;
@@ -49,6 +51,21 @@ pub fn process_req_url(public_path: &str, req_url: &str) -> Result<String> {
         return Ok(req_url[public_path.len() - 1..].to_string());
     }
     Ok(req_url.to_string())
+}
+
+pub(crate) fn get_pkg_name(root: &Path) -> Option<String> {
+    let pkg_json_path = root.join("package.json");
+
+    if pkg_json_path.exists() {
+        let pkg_json = std::fs::read_to_string(pkg_json_path).unwrap();
+        let pkg_json: serde_json::Value = serde_json::from_str(&pkg_json).unwrap();
+
+        pkg_json
+            .get("name")
+            .map(|name| name.as_str().unwrap().to_string())
+    } else {
+        None
+    }
 }
 
 #[cached(key = "String", convert = r#"{ re.to_string() }"#)]
